add KataInfo component
